[x87] Emulate FPATAN using atan2 (#86)
The FPATAN instruction takes in 2 source operands in ST0 and ST1,
corresponding to an X and Y coordinate respectively, and returns the
angle between the positive X axis an (X, Y).

The resulting angle is in the range [-π, +π], which is in line with
`atan2(y, x)`, as opposed to `atan(y / x)` which returns an angle in
[-π/2, +π/2]. This means that any (X, Y) in the 2nd and 3rd quadrants
currently return a wrong angle, potentially breaking 3D math in various
games that use x87 instructions for trigonometry.

Should fix Portal 2.