OpenGL+X11 thunking (#107)
Finished the 64-bit OpenGL thunking implementation, overlaying libGL.so
and libGLX.so to the host libraries. This employs the FEX-Emu strategy
of mapping the X11 Display to a host display and marshalling XVisualInfo
when necessary to avoid thunking the mess that X11 lib is.

Frees up quite a bit of CPU time from the rendering thread in games that
use OpenGL. Games that I've tested include Outlast, Cuphead, Turok,
SuperTuxKart which all seem to benefit.